adding Image component and refactoring Card component
