+++ HCI-lab/src/components/card/style.module.css
@@ -21,3 +21,8 @@
     padding: 2px 16px;
     color: #343131;
   }
+
+  .image {
+    width: 100%;
+    transition: transform 0.2s
+  }
